+++ include/GafferScene/TypeIds.h
@@ -183,6 +183,10 @@ enum TypeId
 	MeshTessellateTypeId = 110639,
 	RenderPassShaderTypeId = 110640,
 	ShaderTweakProxyTypeId = 110641,
+	MergeObjectsTypeId = 110642,
+	MergeMeshesTypeId = 110643,
+	MergePointsTypeId = 110644,
+	MergeCurvesTypeId = 110645,
 
 	PreviewPlaceholderTypeId = 110647,
 	PreviewGeometryTypeId = 110648,