+++ python/GafferSceneUI/__init__.py
@@ -198,6 +198,8 @@ from . import LightPositionToolUI
 from . import MeshTessellateUI
 from . import RenderPassTypeAdaptorUI
 from . import RenderPassShaderUI
+from . import MergeObjectsUI
+from . import MergeMeshesUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.